add squeeze and crop w.i.p

+++ src/imhdlr/tests.rs
@@ -1,7 +1,6 @@
 use super::*;
 use crate::imhdlr::utils::*;
-use std::fs::remove_file;
-use std::fs::{remove_dir_all, File};
+use std::fs::{create_dir_all, remove_dir_all, remove_file, File};
 use std::path::Path;
 
 #[test]
@@ -20,18 +19,61 @@ fn it_imhdlr_get_images() {
 }
 
 #[test]
-fn it_imhdlr_resize_images() {
-    let img1 = "tests/images/dir1/annie-spratt-6wd1f4Zjo_0-unsplash-30x30.jpg";
-    let img2 = "tests/images/dir1/dir2/alex-plesovskich-MHlxTsw5aKY-unsplash-30x30.jpg";
-    remove_file(img1).unwrap();
-    remove_file(img2).unwrap();
-    imhdlr_squeeze("tests/images/", 30, 30, true, true);
+fn it_imhdlr_crop() {
+    let img1 = "tests/images/dir1/annie-spratt-6wd1f4Zjo_0-unsplash-crop-400x400.jpg";
+    let img2 = "tests/images/dir1/dir2/alex-plesovskich-MHlxTsw5aKY-unsplash-crop-400x400.jpg";
+    if file_exists(img1) {
+        remove_file(img1).unwrap();
+    }
+    if file_exists(img2) {
+        remove_file(img2).unwrap();
+    }
+    imhdlr_crop("tests/images/", 400, 400, true);
+    assert!(Path::new(img1).exists(), "File {} does not exist", img1);
+    assert!(Path::new(img2).exists(), "File {} does not exist", img2);
+}
+
+#[test]
+fn it_imhdlr_squeeze() {
+    let img1 = "tests/images/dir1/annie-spratt-6wd1f4Zjo_0-unsplash-400x400.jpg";
+    let img2 = "tests/images/dir1/dir2/alex-plesovskich-MHlxTsw5aKY-unsplash-400x400.jpg";
+    if file_exists(img1) {
+        remove_file(img1).unwrap();
+    }
+    if file_exists(img2) {
+        remove_file(img2).unwrap();
+    }
+    imhdlr_squeeze("tests/images/", 400, 400, true);
+    assert!(Path::new(img1).exists(), "File {} does not exist", img1);
+    assert!(Path::new(img2).exists(), "File {} does not exist", img2);
+}
+
+#[test]
+fn it_imhdlr_squeeze_crop() {
+    let img1 = "tests/images/dir1/annie-spratt-6wd1f4Zjo_0-unsplash-crop-400x400.jpg";
+    let img2 = "tests/images/dir1/dir2/alex-plesovskich-MHlxTsw5aKY-unsplash-crop-400x400.jpg";
+    if file_exists(img1) {
+        remove_file(img1).unwrap();
+    }
+    if file_exists(img2) {
+        remove_file(img2).unwrap();
+    }
+    imhdlr_squeeze_and_crop("tests/images/", 400, 400, true);
     assert!(Path::new(img1).exists(), "File {} does not exist", img1);
     assert!(Path::new(img2).exists(), "File {} does not exist", img2);
 }
 
 fn create_image(file_path: &str) {
     File::create(file_path).expect("Failed to create file");
+}
+
+fn create_dirs(path: &str) {
+    let result = create_dir_all(path);
+    assert!(
+        result.is_ok(),
+        "{}",
+        format!("Directory '{}' already exists or cannot be created.", path)
+    );
 }
 
 fn cleanup(dir: &str) {
@@ -41,4 +83,8 @@ fn cleanup(dir: &str) {
         "{}",
         format!("Directory '{}' does not exist or cannot be deleted.", dir)
     );
+}
+
+fn file_exists(file_path: &str) -> bool {
+    Path::new(file_path).exists()
 }
